Update to 0.1.3

+++ src/main/java/nuclearscience/client/screen/ScreenReactorCore.java
@@ -31,7 +31,7 @@ public class ScreenReactorCore extends GenericContainerScreen<ContainerReactorCo
 	protected void drawGuiContainerForegroundLayer(MatrixStack matrixStack, int mouseX, int mouseY) {
 		super.drawGuiContainerForegroundLayer(matrixStack, mouseX, mouseY);
 		font.func_243248_b(matrixStack, new TranslationTextComponent("gui.reactorcore.deuterium"), titleX, (float) titleY + 14, 4210752);
-		font.func_243248_b(matrixStack, new TranslationTextComponent("gui.reactorcore.temperature", (int) container.getTemperature() / 4 + 15 + "°C"), titleX, (float) titleY + 14 * 3, 4210752);
+		font.func_243248_b(matrixStack, new TranslationTextComponent("gui.reactorcore.temperature", container.getTemperature() / 4 + 15 + "°C"), titleX, (float) titleY + 14 * 3, 4210752);
 		if (container.getTemperature() > TileReactorCore.MELTDOWN_TEMPERATURE_ACTUAL) {
 			if (System.currentTimeMillis() % 1000 < 500) {
 				font.func_243248_b(matrixStack, new TranslationTextComponent("gui.reactorcore.warning"), titleX, (float) titleY + 55, 16711680);
